arranged structure of files
